+++ init.lua
@@ -5,7 +5,7 @@ local BASE = (...):gsub("init$", ""):gsub("([^%.])$", "%1%.")
 local loveshape = {
   _NAME = "loveshape",
   _DESCRIPTION = "Drawable geometric primitives for the LÖVE framework.",
-  _VERSION = "1.0.0",
+  _VERSION = "1.1.0",
   _URL = "https://github.com/binaryfs/loveshape",
   _LICENSE = [[
     MIT License
@@ -31,6 +31,9 @@ local loveshape = {
     SOFTWARE.
   ]],
 }
+
+--- @type loveshape.Ellipse
+loveshape.Ellipse = require(BASE .. "Ellipse")
 
 --- @type loveshape.Circle
 loveshape.Circle = require(BASE .. "Circle")